move CLI to separate file
