+++ index.js
@@ -39,6 +39,7 @@ const fs = require('fs');
 
 // * env
 require('dotenv').config();
+const cliParams = require('./cli');
 
 // * utils
 const transforms = require('./transforms.js');
@@ -253,7 +254,7 @@ async function main(creds = {}, data, opts = {}, isCLI = false, existingConfig) 
 		const envVar = getEnvVars();
 		let cli = {};
 		if (isCLI) {
-			cli = getCLIParams();
+			cli = cliParams();
 			cliData = cli._[0];
 		}
 		config = new importJob({ ...envVar, ...cli, ...creds }, { ...envVar, ...cli, ...opts });
@@ -261,9 +262,19 @@ async function main(creds = {}, data, opts = {}, isCLI = false, existingConfig) 
 
 	if (isCLI) config.verbose = true;
 	const l = logger(config);
-	l(hero.concat(banner));
+	l(cliParams.welcome);
 	global.l = l;
-	
+
+	//for some reason, vscode throws this when --inspect is on...
+	process.on('uncaughtException', (err) => {
+		l(`\nFAILURE!\n\n${err.stack}\n\n${err.message}`);
+	});
+
+
+	process.on('unhandledRejection', (err) => {
+		l(`\nREJECTION!\n\n${err.stack}\n\n${err.message}`);
+	});
+
 	// ETL
 	config.timer.start();
 	const streams = await determineData(data || cliData, config); // always stream[]
@@ -427,137 +438,6 @@ function getEnvVars() {
 	return envCreds;
 }
 
-function getCLIParams() {
-	//! not working!
-	require('@jercle/yargonaut')
-		.style('blue')
-		.helpStyle('green')
-		.errorsStyle('red');
-	const yargs = require('yargs');
-	const args = yargs(process.argv.splice(2))
-		.scriptName("mixpanel-import")
-		.usage(`${hero.concat(banner)}
-		
-usage:\nnpx $0 --yes [file or folder] [options]
-
-ex:
-npx $0 --yes ./events.ndjson --secret 1234 --format jsonl
-npx $0 --yes ./pathToData/ --secret 1234 --format json
-
-DOCS: https://github.com/ak--47/mixpanel-import`)
-		.command('$0', 'import data to mixpanel', () => { })
-		.option("project", {
-			demandOption: false,
-			describe: 'mixpanel project id',
-			type: 'number'
-		})
-		.option("acct", {
-			demandOption: false,
-			describe: 'service account username',
-			type: 'string'
-		})
-		.option("pass", {
-			demandOption: false,
-			describe: 'service account password',
-			type: 'string'
-		})
-		.option("secret", {
-			demandOption: false,
-			describe: 'project API secret (deprecated auth)',
-			type: 'string'
-		})
-		.option("token", {
-			demandOption: false,
-			describe: 'project token',
-			type: 'string'
-		})
-		.option("table", {
-			demandOption: false,
-			describe: "existing lookup table's key",
-			type: 'string'
-		})
-		.option("group", {
-			demandOption: false,
-			describe: 'the group analytics group key',
-			type: 'string'
-		})
-		.option("type", {
-			demandOption: false,
-			alias: "recordType",
-			default: 'event',
-			describe: 'event, user, group, or table',
-			type: 'string'
-		})
-		.option("compress", {
-			demandOption: false,
-			alias: "gzip",
-			default: false,
-			describe: 'gzip on egress',
-			type: 'boolean'
-		})
-		.option("strict", {
-			demandOption: false,
-			default: true,
-			describe: '/import strict mode',
-			type: 'boolean'
-		})
-		.option("logs", {
-			demandOption: false,
-			default: true,
-			describe: 'log import results to file',
-			type: 'boolean'
-		})
-		.option("verbose", {
-			demandOption: false,
-			default: true,
-			describe: 'show progress bar',
-			type: 'boolean'
-		})
-		.option("streamFormat", {
-			demandOption: false,
-			alias: 'format',
-			default: 'jsonl',
-			describe: 'either json or jsonl',
-			type: 'string'
-		})
-		.option("region", {
-			demandOption: false,
-			default: 'US',
-			describe: 'either US or EU',
-			type: 'string'
-		})
-		.option("fixData", {
-			demandOption: false,
-			default: false,
-			describe: 'fix common mistakes',
-			type: 'boolean'
-		})
-		.option("streamSize", {
-			demandOption: false,
-			default: 27,
-			describe: '2^n value of highWaterMark',
-			type: 'number'
-		})
-		.option("recordsPerBatch", {
-			demandOption: false,
-			default: 2000,
-			describe: '# records in each request',
-			type: 'number'
-		})
-		.option("bytesPerBatch", {
-			demandOption: false,
-			default: '2MB',
-			describe: 'max size of each request',
-			type: 'number'
-		})
-		.help()
-		.argv;
-	if (args._.length === 0) {
-		yargs.showHelp();
-		process.exit()
-	}
-	return args;
-}
 
 
 /*
@@ -829,26 +709,9 @@ async function writeLogs(data) {
 	const fileName = `${data.recordType}-import-log-${dateTime}.json`;
 	const filePath = `${fileDir}/${fileName}`;
 	const file = await u.touch(filePath, data, true);
-	l(`\nCOMPLETE\nlog written to: ${file}\n`);
+	l(`\nCOMPLETE\nlog written to:\n${file}`);
 }
 
-/*
------
-WORDS
------
-*/
-
-const hero = String.raw`
-
-            __             ___                  __   __   __  ___ 
-|\/| | \_/ |__)  /\  |\ | |__  |       |  |\/| |__) /  \ |__)  |  
-|  | | / \ |    /~~\ | \| |___ |___    |  |  | |    \__/ |  \  |  
-`;
-const banner = `																  
-... streamer of data... to mixpanel! (v${process.env.npm_package_version || 2})
-  by AK
-  ak@mixpanel.com
-`;
 
 /*
 -------
@@ -860,23 +723,6 @@ const mpImport = module.exports = main;
 mpImport.createMpStream = pipeInterface;
 
 // * this allows the program to run as a CLI
-// todo
 if (require.main === module) {
-	main(undefined, undefined, undefined, true).then(() => { console.log('\nFINISHED!\n'); });
+	main(undefined, undefined, undefined, true).then(() => {  });
 }
-
-/*
----
-CYA
----
-*/
-
-//for some reason, vscode throws this when --inspect is on...
-process.on('uncaughtException', (err) => {
-	l(`\nFAILURE!\n\n${err.stack}\n\n${err.message}`);
-});
-
-
-process.on('unhandledRejection', (err) => {
-	l(`\nREJECTION!\n\n${err.stack}\n\n${err.message}`);
-});
